requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
flake8==3.9.2
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.5.0
mccabe==0.6.1
mockito==1.2.2
multidict==5.1.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.18.0
requests==2.25.1
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
zipp==3.4.1
